refactor: introduces worker struct with identifier
